feat(js): add compatibility mode option to JS compiler
Partially implements #1049.

This required a substantial refactor of `compiler.go`, mostly in order
to be able to test a compiler with different compatibility mode values
during the same test run, which was problematic since `compiler.Compiler`
was a singleton. So the solution is to instead make a new `babel` struct
the singleton, which encapsulates everything needed for Babel to run,
while allowing several `Compiler` instances to exist. In the author's
humble opinion, this has the added benefit of being a cleaner design
(which was further experimentally expanded in 4e52c78a).